llm/pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "google-generativeai>=0.8.5",
    "groq>=0.36.0",
    "huggingface-hub>=0.25.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
